build slink
