+++ internal/go.mod
@@ -3,6 +3,7 @@ module github.com/spectrocloud/CanvOS/internal
 go 1.22.0
 
 require (
+	github.com/sirupsen/logrus v1.9.3
 	github.com/spf13/cobra v1.8.0
 	github.com/twpayne/go-vfs/v5 v5.0.4
 )
@@ -10,5 +11,6 @@ require (
 require (
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/sys v0.17.0 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+	golang.org/x/sys v0.20.0 // indirect
 )
